Implement Diagram.rootElement for BPMNDiagram
Return the value of the feature BPMNDiagram.plane, which redefines
rootElement. In the abstract superclass Diagram, getRootElement now
throws an UnsupportedOperationException.

Implement the opposite reference DiagramElement as well.

This fixes Bug 373642 as far as validation errors are concerned. The
remaining redefined / subsetted features will be implemented in a
separate commit.

Bug: 373642